organize scripts in package
